+++ outputs/sobol_indices_table_Dollar_lost.tex
@@ -1,0 +1,13 @@
+\begin{table}[ht]
+\centering
+\begin{tabular}{lcccccccccccc}
+\hline
+PARAMETER & V_plasma & tau_p_T & tau_p_He3 & n_e_avg & T_e_avg & I_ST & TBR & TBR_DDn & P_aux & eta_th & Cost_per_kWh & SUM \\
+\hline
+RANGE & [1.00e+02,2.05e+03] & [1.00e-03,5.00e+00] & [1.00e-03,5.00e+00] & [1.00e+20,5.00e+20] & [1.00e+01,5.00e+01] & [1.00e+00,3.00e+00] & [5.00e-01,1.30e+00] & [5.00e-01,1.00e+00] & [5.00e+01,5.00e+02] & [2.00e-01,4.00e-01] & [1.50e-01,5.00e-01] & \\
+FIRST-ORDER & -0.0028 & -0.0013 & 0.0000 & 0.0032 & 0.2109 & 0.2598 & 0.0058 & 0.0190 & 0.0000 & 0.1106 & 0.2846 & 0.8900 \\
+TOTAL-ORDER & 0.0010 & 0.0011 & 0.0000 & 0.0014 & 0.2638 & 0.3190 & 0.0059 & 0.0325 & 0.0000 & 0.1426 & 0.3449 & 1.1122 \\
+\hline
+\end{tabular}
+\caption{Sobol indices for the output variable.}
+\end{table}
