+++ outputs/sobol_indices_table_startup_time.tex
@@ -1,0 +1,13 @@
+\begin{table}[ht]
+\centering
+\begin{tabular}{lcccccccccccc}
+\hline
+PARAMETER & V_plasma & tau_p_T & tau_p_He3 & n_e_avg & T_e_avg & I_ST & TBR & TBR_DDn & P_aux & eta_th & Cost_per_kWh & SUM \\
+\hline
+RANGE & [1.00e+02,2.05e+03] & [1.00e-03,5.00e+00] & [1.00e-03,5.00e+00] & [1.00e+20,5.00e+20] & [1.00e+01,5.00e+01] & [1.00e+00,3.00e+00] & [5.00e-01,1.30e+00] & [5.00e-01,1.00e+00] & [5.00e+01,5.00e+02] & [2.00e-01,4.00e-01] & [1.50e-01,5.00e-01] & \\
+FIRST-ORDER & 0.0978 & -0.0001 & 0.0000 & 0.2871 & 0.1507 & 0.0106 & -0.0013 & -0.0017 & 0.0000 & 0.0000 & 0.0000 & 0.5432 \\
+TOTAL-ORDER & 0.5568 & 0.0001 & 0.0000 & 0.5994 & 0.5332 & 0.0682 & 0.0002 & 0.0114 & 0.0000 & 0.0000 & 0.0000 & 1.7692 \\
+\hline
+\end{tabular}
+\caption{Sobol indices for the output variable.}
+\end{table}
